Fix publisher package path to be correct.
Well, that's embarrassing.
