+++ publisher-application/build.gradle
@@ -7,6 +7,7 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter-jersey'
     implementation 'org.springframework.boot:spring-boot-actuator'
+    implementation group: 'io.projectreactor', name: 'reactor-core', version: '3.4.4'
     annotationProcessor 'org.springframework.boot:spring-boot-configuration-processor'
 
     // FHIR
@@ -23,4 +24,8 @@ dependencies {
 
 processResources {
     from("$rootDir/examples")
+}
+
+test {
+    useJUnitPlatform()
 }
