+++ publisher-application/src/main/java/gov/usds/vaccineschedule/publisher/SchedulePublisherApplication.java
@@ -1,4 +1,4 @@
-package cov.usds.vaccineshedule.publisher;
+package gov.usds.vaccineschedule.publisher;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
