+++ publisher-application/src/main/java/gov/usds/vaccineschedule/publisher/respositories/MockSlotRepository.java
@@ -1,8 +1,8 @@
-package cov.usds.vaccineshedule.publisher.respositories;
+package gov.usds.vaccineschedule.publisher.respositories;
 
 import ca.uhn.fhir.context.FhirContext;
 import gov.usds.vaccineschedule.common.helpers.NDJSONToFHIR;
-import org.hl7.fhir.r4.model.Schedule;
+import org.hl7.fhir.r4.model.Slot;
 import org.springframework.stereotype.Service;
 
 import java.io.IOException;
@@ -14,23 +14,24 @@ import java.util.stream.Collectors;
  * Created by nickrobison on 3/26/21
  */
 @Service
-public class MockScheduleRepository implements ScheduleRepository {
+public class MockSlotRepository implements SlotRepository {
 
-    private static final List<Schedule> schedules = createMockSchedules();
+    private static final List<Slot> slots = createMockSlots();
+
 
     @Override
-    public List<Schedule> getAll() {
-        return schedules;
+    public List<Slot> getAll() {
+        return slots;
     }
 
-    private static List<Schedule> createMockSchedules() {
+    private static List<Slot> createMockSlots() {
         final FhirContext fhirContext = FhirContext.forR4();
-        try (InputStream is = MockScheduleRepository.class.getClassLoader().getResourceAsStream("example-schedules.ndjson")) {
+        try (InputStream is = MockScheduleRepository.class.getClassLoader().getResourceAsStream("example-slots.ndjson")) {
 
             final NDJSONToFHIR converter = new NDJSONToFHIR(fhirContext.newJsonParser());
             return converter
                     .inputStreamToResource(is)
-                    .stream().map(r -> (Schedule) r)
+                    .stream().map(r -> (Slot) r)
                     .collect(Collectors.toList());
         } catch (IOException e) {
             throw new RuntimeException("Cannot open file");
