+++ publisher-application/src/main/java/gov/usds/vaccineschedule/publisher/respositories/SlotRepository.java
@@ -1,4 +1,4 @@
-package cov.usds.vaccineshedule.publisher.respositories;
+package gov.usds.vaccineschedule.publisher.respositories;
 
 import org.hl7.fhir.r4.model.Slot;
