+++ publisher-application/src/test/java/gov/usds/vaccineschedule/publisher/PublisherApplicationTests.java
@@ -1,0 +1,16 @@
+package gov.usds.vaccineschedule.publisher;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+/**
+ * Created by nickrobison on 4/6/21
+ */
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public class PublisherApplicationTests {
+
+    @Test
+    public void contextLoads() {
+        // no-op
+    }
+}
